requirements_windows.txt
# Core dependencies for Medical Graph RAG on Windows
# Install with: pip install -r requirements_windows.txt

# Graph and Vector Databases
neo4j==5.23.1
faiss-cpu==1.8.0

# LLM and AI Libraries
google-generativeai==0.6.0
langchain==0.1.14
langchain-community==0.0.31
langchain-core==0.1.40
langchain-text-splitters==0.0.1
sentence-transformers==3.0.1
transformers==4.44.0

# ML/DL Libraries (PyTorch installed separately)
numpy==1.26.4
pandas==2.0.2
scikit-learn==1.5.1
scipy==1.12.0

# Text Processing
nltk==3.8.1
tiktoken==0.7.0
tokenizers==0.19.1

# Data Loading and Processing
datasets==2.19.1
huggingface-hub==0.24.5
python-dotenv==1.0.1

# Document Processing
beautifulsoup4==4.12.2
pdfplumber==0.11.3
python-docx==1.1.2
openpyxl==3.1.5

# Utilities
tqdm==4.66.2
requests==2.31.0
pydantic==2.6.1
